Moved som more general functions to a utils file
